Changed tiles representation
Dropped all graph stuff. Tiles are now located by axial coordinates
(http://www.redblobgames.com/grids/hexagons). It seems there is a
consensus that they're easier to work with. It will certainly help in
case we decide we need to draw the board in the future.
The board is now a dictionary that maps a pair of coordinates to a tile.
Tile is a class which will contain the tile data, like type and number
but, for now, it has only an id.
Added the simple class Vector2 just to help in some cases where we need
to sum and multiply 2-tuples.

Also changed some constructs to support Python2 (I was taking a look at
drawing libraries, and one of the main - PyGame - has only recently
adopted Python3, and you have to compile it from source to obtain
support to it). The code now runs fine with both "python game_loop.py"
or "python3 game_loop.py".